# Fix function-name-case performance