--- a/.changeset/bright-suits-reflect.md
+++ b/.changeset/bright-suits-reflect.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `function-name-case` performance