--- a/lib/rules/function-name-case/index.js
+++ b/lib/rules/function-name-case/index.js
@@ -54,6 +54,8 @@ const rule = (primary, secondaryOptions, context) => {
 		}
 
 		root.walkDecls((decl) => {
+			if (!decl.value.includes('(')) return;
+
 			if (!isStandardSyntaxValue(decl.value)) return;
 
 			let needFix = false;