<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

Ref https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

Before:
Mean: 148.49660561538457 ms
Deviation: 14.86421482337638 ms

After:
Mean: 116.04665178571427 ms
Deviation: 17.283854733889434 ms